backend/requirements.txt
Flask
langchain
langchain-pinecone
langchain-openai
langchain-community
langchain-text-splitters
openai
python-dotenv
pinecone-client
